% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eply_serial}
\alias{eply_serial}
\title{Function \code{eply_serial}}
\usage{
eply_serial(.fun, .expr)
}
\arguments{
\item{.fun}{function to evaluate.}

\item{.expr}{data frame with function calls as rows and function arguments
as columns.}
}
\value{
a list or vector of return values of \code{.fun}.
}
\description{
Serial version of \code{\link{eply}}.
}
\seealso{
\code{\link{eply}}
}

