% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_text}
\alias{eval_text}
\title{Function \code{eval_text}}
\usage{
eval_text(x, .with = environment())
}
\arguments{
\item{x}{character vector encoding unevaluated expressions}

\item{.with}{list, data frame, or environment with the
data accessible to \code{.expr}}
}
\value{
list of evaluated expressions
}
\description{
Evaluate expressions encoded as
elements of a character vector.
}
\seealso{
\code{\link{eply}}
}

