% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals.R
\name{evals}
\alias{evals}
\title{Function \code{evals}}
\usage{
evals(x = NULL, .with = parent.frame(), .simplify = TRUE)
}
\arguments{
\item{x}{character vector of expressions to evaluate}

\item{.with}{list, data frame, or environment with the
data accessible to the expressions in \code{x}}

\item{.simplify}{\code{TRUE} to simplify the result
and \code{FALSE} otherwise}
}
\value{
a list or vector of return values of \code{.fun}.
}
\description{
Evaluate a character vector as a bunch of expressions.
}
\examples{
# Get an example list of supporting data. Could be an environment.
.with <- example.with()
# Row-by-row, evaluate the code in .expr and feed the results to the function.
evals(x = c("a + 1", "b + 2"), .with = .with)
}
\seealso{
\code{\link{eply}}, \code{\link{help_eply}}
}
