% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example.expr}
\alias{example.expr}
\title{\code{example.expr}}
\usage{
example.expr()
}
\value{
Example \code{.expr} argument to \code{\link{eply}}.
}
\description{
Return example \code{.expr} argument for \code{\link{eply}}.
}
\examples{
#' Get an example .expr argument to eply().
#' See the examples of the eply() function for usage.
example.expr()
}
\seealso{
\code{\link{eply}}
}
