% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example.fun}
\alias{example.fun}
\title{\code{example.fun}}
\usage{
example.fun(x, y)
}
\arguments{
\item{x}{numeric argument}

\item{y}{nonzero numeric argument}
}
\value{
Example \code{.fun} argument to \code{\link{eply}}.
}
\description{
Example \code{.fun} argument to \code{\link{eply}}.
}
\seealso{
\code{\link{eply}}
}
