% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_omniply}
\alias{example_omniply}
\title{Function \code{example_omniply}}
\usage{
example_omniply(example = list_examples_omniply())
}
\arguments{
\item{example}{name of the example. To see all the available example names, 
run \code{\link{list_examples_omniply}}.}
}
\description{
Copy a omniply example to the current working directory.
To see the names of all the examples, run \code{\link{list_examples_omniply}}.
}
\seealso{
\code{\link{list_examples_omniply}}
}

