% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{help_eply}
\alias{help_eply}
\title{Function \code{help_eply}}
\usage{
help_eply()
}
\description{
Prints links for tutorials, troubleshooting, bug reports, etc.
}
\examples{
help_eply()
}
\seealso{
\code{\link{eply}}, \code{\link{evals}}
}
