% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_serial}
\alias{is_serial}
\title{Function \code{is_serial}}
\usage{
is_serial(.expr, .split = NULL, .tasks = 1, .os = Sys.info()[["sysname"]])
}
\arguments{
\item{.expr}{argument to \code{\link{eply}}}

\item{.split}{argument to \code{\link{eply}}}

\item{.tasks}{argument to \code{\link{eply}}}

\item{.os}{operating system}
}
\value{
\code{TRUE} if \code{\link{eply}} will run serially
and \code{FALSE} if parallelized.
}
\description{
Determine whether to execute \code{\link{eply}} serially.
}
\seealso{
\code{\link{eply}}
}

