% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{list_examples_omniply}
\alias{list_examples_omniply}
\title{Function \code{list_examples_omniply}}
\usage{
list_examples_omniply()
}
\value{
a names of all the omniply examples.
}
\description{
Return the names of all the omniply examples.
}
\seealso{
\code{\link{example_omniply}}
}

