% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meval.R
\name{meval}
\alias{meval}
\title{Function \code{meval}}
\usage{
meval(x = NULL, .with = environment(), .tasks = 1)
}
\arguments{
\item{x}{character vector of quoted expressions}

\item{.with}{list, data frame, or environment with the
data accessible to \code{x}}

\item{.tasks}{number of parallel tasks}
}
\value{
results of evaluating \code{x} elementwise
(array or vector if \code{.simplify} is \code{TRUE} 
and a list otherwise)
}
\description{
evaluate a character vector of quoted expressions 
and return a collection of results
}
\seealso{
\code{\link{meval}}, \code{\link{eply}}, \code{\link{help_eply}}
}

