% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quotes}
\alias{quotes}
\title{Function \code{quotes}}
\usage{
quotes(x = NULL)
}
\arguments{
\item{x}{character vector or object to be coerced to character.}
}
\value{
character vector with quotes around it
}
\description{
Put quotes around each element of a character vector.
}
\seealso{
\code{\link{eply}}
}

