% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{seval}
\alias{seval}
\title{Function \code{seval}}
\usage{
seval(x, .with = environment())
}
\arguments{
\item{x}{character vector of quoted expressions}

\item{.with}{list, data frame, or environment with the
data accessible to \code{x}}
}
\value{
results of evaluating \code{x} elementwise
}
\description{
evaluate a character vector of quoted expressions 
and return a SIMPLIFIED list of results
}
\seealso{
\code{\link{veval}}, \code{\link{eply}}, \code{\link{help_eply}}
}

