% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{strings}
\alias{strings}
\title{Function \code{strings}}
\usage{
strings(...)
}
\arguments{
\item{...}{unquoted symbols to turn into character strings.}
}
\value{
a character vector
}
\description{
Turn valid expressions into character strings.
}
\seealso{
\code{\link{quotes}}, \code{\link{unquote}},
\code{\link{eply}}, \code{\link{help_eply}}
}
