% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{unquote}
\alias{unquote}
\title{Function \code{unquote}}
\usage{
unquote(x = NULL)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without leading or trailing escaped quotes around it
}
\description{
Remove leading and trailing escaped quotes from character strings.
}
\seealso{
\code{\link{eply}}
}

