% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{veval}
\alias{veval}
\title{Function \code{veval}}
\usage{
veval(x, .with = environment())
}
\arguments{
\item{x}{character vector of quoted expressions}

\item{.with}{list, data frame, or environment with the
data accessible to \code{x}}
}
\value{
results of evaluating \code{x} elementwise
}
\description{
evaluate a character vector of 
quoted expressions and return a list of results
}
\seealso{
\code{\link{seval}} \code{\link{eply}}, \code{\link{help_eply}}
}

