% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eply.R
\name{eval_text}
\alias{eval_text}
\title{Function \code{eval_text}}
\usage{
eval_text(x)
}
\arguments{
\item{x}{character vector encoding unevaluated expressions}
}
\value{
list of evaluated expressions
}
\description{
Evaluate expressions encoded as
elements of a character vector.
}

