% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example.expr}
\alias{example.expr}
\title{\code{example.expr}}
\usage{
example.expr()
}
\value{
Example \code{.expr} argument to \code{\link{eply}}.
}
\description{
Example \code{.expr} argument to \code{\link{eply}}.
}
\seealso{
\code{\link{eply}}
}

