% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example.fun}
\alias{example.fun}
\title{\code{example.fun}}
\usage{
example.fun(x, y)
}
\arguments{
\item{x}{numeric argument}

\item{y}{nonzero numeric argument}
}
\value{
Example \code{.fun} argument to \code{\link{eply}}.
}
\description{
Example \code{.fun} argument to \code{\link{eply}}.
}
\examples{
#' Get an example .fun argument to eply().
#' See the examples of the eply() function for usage.
example.fun
example.fun(x = c(4, 2), y = c(2, 2))
}
\seealso{
\code{\link{eply}}
}
