% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example.with}
\alias{example.with}
\title{\code{example.with}}
\usage{
example.with()
}
\value{
example \code{.with} argument of \code{\link{eply}}
}
\description{
Return example \code{.with} argument of \code{\link{eply}}.
}
\seealso{
\code{\link{eply}}
}
