% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{quotes}
\alias{quotes}
\title{Function \code{quotes}}
\usage{
quotes(x = NULL, single = FALSE)
}
\arguments{
\item{x}{character vector or object to be coerced to character.}

\item{single}{Add single quotes if \code{TRUE} and double quotes otherwise.}
}
\value{
character vector with quotes around it
}
\description{
Put quotes around each element of a character vector.
}
\seealso{
\code{\link{unquote}}, \code{\link{strings}},
\code{\link{eply}}, \code{\link{help_eply}}
}

